debugging

+++ container_scripts/containers/60_swag/container/cleanup.sh
@@ -6,4 +6,9 @@ then
    exit 1
 fi
 
-runuser --login ${USER_NAME} -c "podman volume rm ${SWAG_VOL_NAME}"
+if [[ ${DEBUG} == "FALSE" ]];
+then
+   runuser --login ${USER_NAME} -c "podman volume rm ${SWAG_VOL_NAME}"
+
+   rm dockerfiles/swag/default
+fi
